+++ src/cercis/__init__.py
@@ -39,13 +39,17 @@ from blib2to3.pytree import Leaf, Node
 from cercis.cache import Cache, get_cache_info, read_cache, write_cache
 from cercis.comments import normalize_fmt_off
 from cercis.const import (
+    DEFAULT_BASE_INDENTATION_SPACES,
     DEFAULT_CLOSING_BRACKET_EXTRA_INDENT,
     DEFAULT_COLLAPSE_NESTED_BRACKETS,
     DEFAULT_EXCLUDES,
     DEFAULT_FUNCTION_DEFINITION_EXTRA_INDENT,
     DEFAULT_INCLUDES,
     DEFAULT_LINE_LENGTH,
+    DEFAULT_OTHER_LINE_CONTINUATION_EXTRA_INDENT,
     DEFAULT_SINGLE_QUOTE,
+    DEFAULT_TAB_WIDTH,
+    DEFAULT_USE_TABS,
     DEFAULT_WRAP_LINE_WITH_LONG_STRING,
     DEFAULT_WRAP_PRAGMA_COMMENTS,
     STDIN_PLACEHOLDER,
@@ -203,6 +207,17 @@ def validate_regex(
         raise click.BadParameter(f"Not a valid regular expression: {e}") from None
 
 
+def validate_positive_integer(
+        ctx: click.Context,
+        param: click.Parameter,
+        value: Optional[int],
+) -> Optional[int]:
+    if not isinstance(value, int) or value <= 0:
+        raise click.BadParameter("Not a positive integer") from None
+
+    return value
+
+
 @click.command(
     context_settings={"help_option_names": ["-h", "--help"]},
     # While Click does set this field automatically using the docstring, mypyc
@@ -219,6 +234,40 @@ def validate_regex(
     show_default=True,
 )
 @click.option(
+    "-tab",
+    "--use-tabs",
+    type=bool,
+    show_default=True,
+    default=DEFAULT_USE_TABS,
+    help="If True, use tabs as indentation, rather than spaces.",
+)
+@click.option(
+    "-tw",
+    "--tab-width",
+    type=int,
+    show_default=True,
+    default=DEFAULT_TAB_WIDTH,
+    callback=validate_positive_integer,
+    help=(
+        "The tab width to assume when calculating line length."
+        " This option is only effective when --use-tabs is set to True."
+        " Use a value that matches your editor's tab size setting for"
+        " consistent behaviors."
+    ),
+)
+@click.option(
+    "-bis",
+    "--base-indentation-spaces",
+    type=int,
+    show_default=True,
+    default=DEFAULT_BASE_INDENTATION_SPACES,
+    callback=validate_positive_integer,
+    help=(
+        "How many spaces to use for each indentation level."
+        " It has no effect when --use-tabs is True."
+    ),
+)
+@click.option(
     "-fdei",
     "--function-definition-extra-indent",
     type=bool,
@@ -228,6 +277,14 @@ def validate_regex(
         "If True, use 8 spaces as indent in function definition;"
         " otherwise, use 8 (Black's default)."
     ),
+)
+@click.option(
+    "-olcei",
+    "--other-line-continuation-extra-indent",
+    type=bool,
+    show_default=True,
+    default=DEFAULT_OTHER_LINE_CONTINUATION_EXTRA_INDENT,
+    help="If True, add an extra indentation level when wrapping longer lines.",
 )
 @click.option(
     "-cbei",
@@ -506,6 +563,10 @@ def main(  # noqa: C901
         wrap_line_with_long_string: bool,
         collapse_nested_brackets: bool,
         wrap_pragma_comments: bool,
+        base_indentation_spaces: int,
+        other_line_continuation_extra_indent: bool,
+        use_tabs: bool,
+        tab_width: int,
         target_version: List[TargetVersion],
         check: bool,
         diff: bool,
@@ -634,6 +695,10 @@ def main(  # noqa: C901
         wrap_line_with_long_string=wrap_line_with_long_string,
         collapse_nested_brackets=collapse_nested_brackets,
         wrap_pragma_comments=wrap_pragma_comments,
+        base_indentation_spaces=base_indentation_spaces,
+        other_line_continuation_extra_indent=other_line_continuation_extra_indent,
+        use_tabs=use_tabs,
+        tab_width=tab_width,
     )
 
     if code is not None: