+++ src/cercis/mode.py
@@ -18,11 +18,15 @@ else:
     from typing import Final
 
 from cercis.const import (
+    DEFAULT_BASE_INDENTATION_SPACES,
     DEFAULT_CLOSING_BRACKET_EXTRA_INDENT,
     DEFAULT_COLLAPSE_NESTED_BRACKETS,
     DEFAULT_FUNCTION_DEFINITION_EXTRA_INDENT,
     DEFAULT_LINE_LENGTH,
+    DEFAULT_OTHER_LINE_CONTINUATION_EXTRA_INDENT,
     DEFAULT_SINGLE_QUOTE,
+    DEFAULT_TAB_WIDTH,
+    DEFAULT_USE_TABS,
     DEFAULT_WRAP_LINE_WITH_LONG_STRING,
     DEFAULT_WRAP_PRAGMA_COMMENTS,
 )
@@ -198,6 +202,12 @@ class Mode:
     wrap_line_with_long_string: bool = DEFAULT_WRAP_LINE_WITH_LONG_STRING
     collapse_nested_brackets: bool = DEFAULT_COLLAPSE_NESTED_BRACKETS
     wrap_pragma_comments: bool = DEFAULT_WRAP_PRAGMA_COMMENTS
+    base_indentation_spaces: int = DEFAULT_BASE_INDENTATION_SPACES
+    other_line_continuation_extra_indent: bool = (
+        DEFAULT_OTHER_LINE_CONTINUATION_EXTRA_INDENT
+    )
+    use_tabs: bool = DEFAULT_USE_TABS
+    tab_width: int = DEFAULT_TAB_WIDTH
 
     def __post_init__(self) -> None:
         if self.experimental_string_processing: