+++ src/cercis/trans.py
@@ -23,6 +23,8 @@ from typing import (
     Union,
 )
 
+from cercis.indent import Indent
+
 if sys.version_info < (3, 8):
     from typing_extensions import Final, Literal
 else:
@@ -1101,7 +1103,7 @@ class BaseStringSplitter(StringTransformer):
         #   NN: The leaf that is after N.
 
         # WMA4 the whitespace at the beginning of the line.
-        offset = line.depth * 4
+        offset: int = line.calc_total_indent_width()
 
         if is_valid_index(string_idx - 1):
             p_idx = string_idx - 1
@@ -1455,7 +1457,7 @@ class StringSplitter(BaseStringSplitter, CustomSplitMapMixin):
                 characters expand to two columns).
             """
             result = self.line_length
-            result -= line.depth * 4
+            result -= line.calc_total_indent_width()
             result -= 1 if ends_with_comma else 0
             result -= string_op_leaves_length
             return result
@@ -1466,11 +1468,11 @@ class StringSplitter(BaseStringSplitter, CustomSplitMapMixin):
         # The last index of a string of length N is N-1.
         max_break_width -= 1
         # Leading whitespace is not present in the string value (e.g. Leaf.value).
-        max_break_width -= line.depth * 4
+        max_break_width -= line.calc_total_indent_width()
         if max_break_width < 0:
             yield TErr(
                 f"Unable to split {LL[string_idx].value} at such high of a line depth:"
-                f" {line.depth}"
+                f" {line.calc_total_indent_width()}"
             )
             return
 
@@ -1893,7 +1895,9 @@ class StringParenWrapper(BaseStringSplitter, CustomSplitMapMixin):
                 char == " " or char in SPLIT_SAFE_CHARS for char in string_value
             ):
                 # And will still violate the line length limit when split...
-                max_string_width = self.line_length - ((line.depth + 1) * 4)
+                max_string_width: int = self.line_length - (
+                    (line.calc_total_indent_width() + 1) * 4
+                )
                 if str_width(string_value) > max_string_width:
                     # And has no associated custom splits...
                     if not self.has_custom_splits(string_value):
@@ -2139,7 +2143,7 @@ class StringParenWrapper(BaseStringSplitter, CustomSplitMapMixin):
         string_value = LL[string_idx].value
         string_line = Line(
             mode=line.mode,
-            depth=line.depth + 1,
+            depth=line.depth + (Indent.OTHER_LINE_CONTINUATION,),
             inside_brackets=True,
             should_split_rhs=line.should_split_rhs,
             magic_trailing_comma=line.magic_trailing_comma,